refactor: split transcription.rs into transcription/ module (SRP)
